Move psQuery -> lib/ps.query()
