+++ index.js
@@ -1,47 +1,7 @@
 var ChildProcess = require('child_process');
 var csv = require('csv');
 var libExec = require('./lib/exec');
-
-function psQuery(args, callback){
-
-    var cmd = 'ps ' + args;
-    libExec.runCommand(cmd, function (err, stdout) {
-
-        if (err) {
-            return callback(err);
-        }
-
-        var results = [];
-        stdout.split('\n').map(function (row) {
-
-            var matches = row.match(/(\d+) (.*)/);
-            if (!matches) {
-                return;
-            }
-
-            var pid = parseInt(matches[1], 10);
-            var whole = matches[2];
-            var slashIndex = whole.lastIndexOf('/');
-            var command = whole.substr(slashIndex + 1);
-            var execDir = whole.substr(0, slashIndex);
-
-            // 123, Steam, /opts/homebrew/casks/steam.app/contents/MacOS/
-            results.push({
-                pid: pid,
-                command: command,
-                execDir: execDir
-            });
-        });
-
-        //filter our commands that we don't know where they're from
-        results = results.filter(function (result){
-
-            return result.execDir.length > 0;
-        });
-
-        callback(null, results);
-    });
-};
+var libPs = require('./lib/ps');
 
 function wmicQuery(args, callback){
 
@@ -96,7 +56,7 @@ module.exports.list = function (callback){
         wmicQuery('', callback);
     } else {
         // OS X/Linux check
-        psQuery('-A -o pid,comm', callback);
+        libPs.query('-A -o pid,comm', callback);
     }
 };
 
@@ -107,7 +67,7 @@ module.exports.lookup = function (pid, callback){
         wmicQuery('where processid=' + pid, callback);
     } else {
         // OS X/Linux check
-        psQuery('-o pid,comm -p ' + pid, callback);
+        libPs.query('-o pid,comm -p ' + pid, callback);
     }
 };
 
@@ -118,6 +78,6 @@ module.exports.detailedLookup = function (pid, callback){
         wmicQuery('where processid=' + pid, callback);
     } else {
         // OS X/Linux check
-        psQuery('-o pid,args -p ' + pid, callback);
+        libPs.query('-o pid,args -p ' + pid, callback);
     }
 };
